Created update.py and RESTupdate.py
Currently, only RESTupdate.py is working. update.py is still in progress.
RESTupdate.py currently can only check the local version of the script or file to be updated
and checks the version of the script or file on the server. If the server version is newer, it will download the new version and replace the old version.
There are still issues with deleting the temporary directory and files after the update is complete.